requirements.txt
Flask-API.yandex
aniso8601==3.0.2
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
cryptography==2.1.4
cycler==0.10.0
dnspython==1.15.0
enum34==1.1.6
firebase-admin==2.13.0
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-RESTful==0.3.6
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.12.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.9.0
idna==2.6
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
MarkupSafe==1.0
marshmallow==2.15.5
matplotlib==2.2.3
msgpack==0.5.6
numpy==1.15.1
pandas==0.23.4
pipenv==2018.7.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.4.3
pyparsing==2.2.0
Pyrebase==3.0.27
python-dateutil==2.7.3
pytz==2018.5
pyxdg==0.25
requests==2.19.1
rsa==4.0
SecretStorage==2.3.1
six==1.11.0
subprocess32==3.5.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
